Some bug fixes

+++ common/GlobalDefs.h
@@ -20,7 +20,7 @@ typedef struct _SymStore{
     int fileIndex;
     char *sym;
     Elf64_Sym sym_info;
-    int outIndex;
+    uint64_t outIndex;
 }SymStore;
 
 #define GetSymStructFromSymSet(str,isDynamic) ((SymStore *)(Set_Find(GetCurSymSet(isDynamic),str)))
